travaia-ai/backend/document-report-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.18.4
firebase-admin==6.2.0
pydantic==2.5.0
structlog==23.2.0
python-multipart==0.0.6
httpx==0.25.2
reportlab==4.0.7
Pillow==10.1.0
python-dotenv==1.0.0
slowapi==0.1.9
tenacity==8.2.3
pypdf2==3.0.1
aiofiles==23.2.1
requests==2.31.0
PyJWT==2.8.0
redis==5.0.1